Move Module to core
